strategy/requirements.txt
psycopg2-binary==2.9.9
redis==5.0.4
pandas==2.2.2
python-dotenv==1.0.1
